[Auth] - Authorize role & permisson for user

+++ app/Http/Livewire/Admin/Categories/AdminCategories.php
@@ -9,27 +9,25 @@ use WireUi\Traits\Actions;
 use Livewire\WithPagination;
 use App\Traits\tableSortingTrait;
 
-class AdminCategories extends Component
-{
+class AdminCategories extends Component {
 
     use WithPagination;
     use tableSortingTrait;
     use Actions;
 
     public $categoryId;
-    public $search = '';
+    public $search              = '';
     public $selectAllCategories = false;
-    public $selectedCategories = [];
-    protected $queryString = [
+    public $selectedCategories  = [];
+    protected $queryString      = [
         'search' => ['except' => '']
     ];
     protected $listeners = [
         'resetSelected' => 'resetSelected',
-        'editCategory' => 'editCategory'
+        'editCategory'  => 'editCategory'
     ];
 
-    public function editCategory($categoryId)
-    {
+    public function editCategory($categoryId) {
         $this->emitTo(
             'admin.categories.edit-categories',
             'editCategory',
@@ -37,17 +35,24 @@ class AdminCategories extends Component
         );
     }
 
-    public function destroyCategory()
-    {
+    public function destroyCategory() {
         try {
-            $category = Category::findOrFail($this->categoryId);
-            $categoryName = $category->name;
-            $category->delete();
+            try {
+                $this->authorize('delete_category');
+                $category     = Category::findOrFail($this->categoryId);
+                $categoryName = $category->name;
+                $category->delete();
 
-            $this->notification()->success(
-                $title = 'Đã xóa !!!',
-                $description = 'Đã xóa danh mục <strong>' . $categoryName . '</strong>'
-            );
+                $this->notification()->success(
+                    $title = 'Đã xóa !!!',
+                    $description = 'Đã xóa danh mục <strong>' . $categoryName . '</strong>'
+                );
+            } catch (Exception $e) {
+                $this->notification()->error(
+                    $title = 'Đã xảy ra lỗi !!!',
+                    $description = 'Bạn không có quyền xóa danh mục'
+                );
+            }
         } catch (Exception $e) {
             $this->notification()->error(
                 $title = 'Đã xảy ra lỗi !!!',
@@ -56,39 +61,56 @@ class AdminCategories extends Component
         }
     }
 
-    public function deleteCategory($categoryId)
-    {
+    public function deleteCategory($categoryId) {
         $this->categoryId = $categoryId;
     }
 
-    public function deleteSelected()
-    {
-        Category::whereIn('id', $this->selectedCategories)->delete();
-        $this->resetSelected();
+    public function deleteSelected() {
+        try {
+            try {
+                $this->authorize('delete_category');
+                $categories = Category::whereIn('id', $this->selectedCategories)->get();
+                foreach ($categories as $category) {
+                    $this->deleteImage($category->thumbnail);
+                    $category->delete();
+                }
+
+                $this->notification()->success(
+                    $title = 'Đã xóa !!!',
+                    $description = 'Đã xóa các danh mục đã chọn'
+                );
+            } catch (Exception $e) {
+                $this->notification()->error(
+                    $title = 'Đã xảy ra lỗi !!!',
+                    $description = 'Bạn không có quyền xóa danh mục'
+                );
+            }
+        } catch (Exception $e) {
+            $this->notification()->error(
+                $title = 'Đã xảy ra lỗi !!!',
+                $description = 'Xóa danh mục thất bại'
+            );
+        }
     }
 
-    public function resetSelected()
-    {
+    public function resetSelected() {
         $this->selectAllCategories = false;
-        $this->selectedCategories = [];
+        $this->selectedCategories  = [];
     }
 
-    public function updatedSelectAllCategories($value)
-    {
+    public function updatedSelectAllCategories($value) {
         if ($value) {
-            $this->selectedCategories = $this->getCategories()->pluck('id')->map(fn($id) => (string)$id);
+            $this->selectedCategories = $this->getCategories()->pluck('id')->map(fn($id) => (string) $id);
         } else {
             $this->selectedCategories = [];
         }
     }
 
-    public function getCategories()
-    {
+    public function getCategories() {
         return Category::all();
     }
 
-    public function updatedSelectedCategories($value)
-    {
+    public function updatedSelectedCategories($value) {
         if ($value && count($value) === $this->getCategories()->count()) {
             $this->selectAllCategories = true;
         } else {
@@ -96,15 +118,14 @@ class AdminCategories extends Component
         }
     }
 
-    public function render()
-    {
+    public function render() {
         $categories = Category::select('categories.*')
             ->when($this->search, function ($query) {
                 $query->where('categories.name', 'like', '%' . $this->search . '%');
             })
-            ->orderBy($this->sortColumnName, $this->sortDirection)
-            ->orderBy($this->sortByColumn(), $this->sortDirection())
-            ->paginate($this->perPage);
+                     ->orderBy($this->sortColumnName, $this->sortDirection)
+                     ->orderBy($this->sortByColumn(), $this->sortDirection())
+                     ->paginate($this->perPage);
 
         return view('livewire.admin.categories.admin-categories', compact('categories'));
     }
